requirements.txt
Flask==2.3.2
python-dotenv==1.0.0
json5==0.9.11
slugify
